krz/requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.10.0
bleach==4.1.0
cffi==1.15.0
colorama==0.4.4
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
Django==4.0.3
entrypoints==0.4
executing==0.8.3
ipykernel==6.9.2
ipython==8.1.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
MarkupSafe==2.1.0
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.2.0
nest-asyncio==1.5.4
notebook==6.4.9
numpy==1.22.3
packaging==21.3
pandas==1.4.1
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
pip==21.1.1
prometheus-client==0.13.1
prompt-toolkit==3.0.28
psutil==5.9.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.11.2
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2021.3
pywin32==303
pywinpty==2.0.5
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.1
scikit-learn==1.0.2
scipy==1.8.0
Send2Trash==1.8.0
setuptools==56.0.0
six==1.16.0
sklearn==0.0
soupsieve==2.3.1
sqlparse==0.4.2
stack-data==0.2.0
terminado==0.13.3
testpath==0.6.0
threadpoolctl==3.1.0
tornado==6.1
traitlets==5.1.1
tzdata==2021.5
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.2
django-import-export==2.8.0